JSPEC

Documentation for JSPEC.

JSPEC.CreateDataSet

JSPEC.FindRebinSchema

JSPEC.GenRebin

JSPEC.GetKnownInstruments

JSPEC.IgnoreChannels

JSPEC.ImportData

JSPEC.ImportOtherData

JSPEC.PlotRaw

JSPEC.RebinData

JSPEC.CreateDataSet — Method

CreateDataSet(Name::String, Instrument::String; verbose=true)::Dict

Create JSPECDataSent entry. 'Name' is the arbitrary name of the dataset one may choose and 'Instrument' is one of supperted instrument by the package. If 'verbose' is set, it generates, if needed, a warning message if data are now properly processed.

Examples

newdataset = CreateDataSet("XRTTest","Swift-XRT")

# output

Dict{Any, Any} with 3 entries:
  "Name"       => "XRTTest"
  "Instrument" => "Swift-XRT"
  "Created"    => true

JSPEC.FindRebinSchema — Method

FindRebinSchema(x,ey;minSN=5)::AbstractVector{Real}

Compute the rebin schema to guarantee that the S/N is at least 'minSN' in each bin (or channel). 'x' and 'ex' are the input data and relative uncertainties.

Examples

x = [1.,2.,3.,4.,]
ex = [0.1,0.5,0.6,0.05]

JSPEC.FindRebinSchema(x,ex)

# output

3-element Vector{Real}:
 1
 3
 4

JSPEC.GenRebin — Method

GenRebin(x,rebs)::AbstractVector{Real}

Rebin input data following a given rebin schema.

Examples

x = [1.,2.,3.,4.,]
rbs = [1,3,4]

JSPEC.GenRebin(x,rbs)

# output

3-element Vector{Real}:
 1.0
 2.5
 4.0

JSPEC.GetKnownInstruments — Method

GetKnownInstruments()

Returns the instruments currently supported by the JSPEC package.

Examples

@show GetKnownInstruments()

JSPEC.IgnoreChannels — Method

IgnoreChannels(ds:Dict, chns; verbose=true)

Ignore channels in the input data. 'ds' is the JSPEC data set dictionary, 'chns' is a vector of channels to be ignored, e.g. [0,1,2,3] or even [0:4, 1000:1023]. Please remeber that channel numebering starts with 0. If 'verbose' is set, it generates, if needed, a warning message if data are now properly processed.

Examples

IgnoreChannels(newdataset,[0,1,2,3])

JSPEC.ImportData — Method

ImportData(ds::Dict; rmffile::String="", arffile::String="", srcfile::String="", bckfile::String="", verbose=true)

Import data from "multi-channel" instruments (e.g., Swift-XRT) as FITS files and add fields to the JSPECDataSet dictionary. 'ds' is the JSPEC data set dictionary, 'rmfile' is the RMF response matrix, 'arffile' the effective area matrix, 'srcfile' the counts (or rates) for the source, and 'bckfile' the counts or rates for the background. If 'verbose' is set, it generates, if needed, a warning message if data are now properly processed.

Examples

fnrmf = joinpath("data","wt.rmf")
fnarf = joinpath("data","wt.arf")
fnpisrc = joinpath("data","wtsource.pi")
fnpibck = joinpath("data","tback.pi");

ImportData(newdataset, rmffile=fnrmf,arffile=fnarf,srcfile=fnpisrc,bckfile=fnpibck)

JSPEC.ImportOtherData — Method

ImportOtherData(ds::Dict, energy, phflux, ephflux; bandwidth=1., verbose=true)

Import data already in physical units, 'ds' is a JEPC dictionary, 'energy' is KeV, 'phflux' is the corresponding photon flux density in photons cm^-2 s^-1 KeV^-1 and 'ephflux' the uncertainty. In case it is a photon flux (photons cm^-2 s^-1) the with, in KeV, of the band, 'bandwidth' must be provided. If 'verbose' is set, it generates, if needed, a warning message if data are now properly processed.

Examples

ImportOtherData(newdataset, [1.,2.,3.,4], [0.1,0.2,0.3,0.4], [0.01,0.02,0.03,0.04])

JSPEC.PlotRaw — Method

PlotRaw(ds:Dict; xlbl="Channels", ylbl="Counts", tlbl=ds.Name, verbose=true)::Figure

Draw a plot of the raw input data. 'ds' is the JSPEC data set dictionary, 'xlbl' and 'ylbl' are the labels for the x and y axes, whule 'tlbl' is the plot title. If 'verbose' is set, it generates, if needed, a warning message if data are now properly processed.

Examples

figraw = PlotRaw(newdataset)

JSPEC.RebinData — Method

RebinData(ds::Dict;minSN=5,verbose=true)

Rebin input data with a given mininum S/N per bin. If 'verbose' is set, it generates, if needed, a warning message if data are now properly processed.

Examples

RebinData(newdataset)