JSPEC

Documentation for JSPEC.

JSPEC.CreateDataSet

JSPEC.FindRebinSchema

JSPEC.GenRebin

JSPEC.GenResponseMatrix

JSPEC.GetKnownInstruments

JSPEC.IgnoreChannels

JSPEC.ImportData

JSPEC.ImportOtherData

JSPEC.PlotRaw

JSPEC.PlotRebinned

JSPEC.RebinAncillaryData

JSPEC.RebinData

JSPEC.CreateDataSet — Method

CreateDataSet(Name::String, Instrument::String; verbose=true)::Dict

Create a JSPECDataSent entry.

#Arguments

Name is the arbitrary name of the dataset.

Instrument is one of supperted instrument by the package.

verbose enables warning messages.

Examples

newdataset = CreateDataSet("XRTTest","Swift-XRT")

# output

Dict{Any, Any} with 3 entries:
  "Name"       => "XRTTest"
  "Instrument" => "Swift-XRT"
  "Created"    => true

JSPEC.FindRebinSchema — Method

FindRebinSchema(x,ey;minSN=5)::AbstractVector{Real}

Compute the rebin schema to guarantee that the S/N is at least minSN in each bin (or channel).

Arguments

x input array.

ex uncertainties.

Examples

x = [1.,2.,3.,4.,]
ex = [0.1,0.5,0.6,0.05]

JSPEC.FindRebinSchema(x,ex)

# output

3-element Vector{Real}:
 1
 3
 4

JSPEC.GenRebin — Method

GenRebin(x,rebs)::AbstractVector{Real}

Rebin input data following a given rebin schema.

Arguments

x input array.

rebs array with rebin schema.

Examples

x = [1.,2.,3.,4.,]
rbs = [1,3,4]

JSPEC.GenRebin(x,rbs)

# output

3-element Vector{Real}:
 1.0
 2.5
 4.0

JSPEC.GenResponseMatrix — Method

GenResponseMatrix(ds::Dict; verbose=true)

Generate a rebinned response matrix following the rebin schema identified for input data.

Arguments

ds JSPEC data set dictionary.

Examples

GenResponseMatrix(newdataset)

JSPEC.GetKnownInstruments — Method

GetKnownInstruments()

Returns the instruments currently supported by the JSPEC package.

Examples

@show GetKnownInstruments()

JSPEC.IgnoreChannels — Method

IgnoreChannels(ds:Dict, chns; verbose=true)

Ignore channels in the input data.

Arguments

ds JSPEC data set dictionary.

chns vector of channels to be ignored, e.g. [0,1,2,3] or even [0:4, 1000:1023]. Pay attention that channel numbering starts at 0.

verbose enables warning messages.

Examples

IgnoreChannels(newdataset,[0,1,2,3])

JSPEC.ImportData — Method

ImportData(ds::Dict; rmffile::String="", arffile::String="", srcfile::String="", bckfile::String="", verbose=true)

Import data from "multi-channel" instruments (e.g., Swift-XRT).

Arguments

ds` JSPEC data set dictionary.

rmfile` RMF response matrix.

arffile effective area matrix.

srcfile source counts (or rate).

bckfile background counts (or rate).

verbose enables warning messages.

Examples

fnrmf = joinpath("data","wt.rmf")
fnarf = joinpath("data","wt.arf")
fnpisrc = joinpath("data","wtsource.pi")
fnpibck = joinpath("data","tback.pi");

ImportData(newdataset, rmffile=fnrmf,arffile=fnarf,srcfile=fnpisrc,bckfile=fnpibck)

JSPEC.ImportOtherData — Method

ImportOtherData(ds::Dict, energy, phflux, ephflux; bandwidth=1., verbose=true)

Import data already in physical units.

Arguments

ds JSPC dictionary.

energy input energy (KeV).

phflux photon flux density ($photons~cm{^-2}~s{^-1}~KeV{^-1})$.

ephflux photon flux density uncertainty.

bandwidth band width (KeV).

`verbose? enable warning message.

Bandwidth is needed only in case photon flux ($photons~cm{^-2}~s{^-1})$, rather then photon flux density, is provided.

Examples

ImportOtherData(newdataset, [1.,2.,3.,4], [0.1,0.2,0.3,0.4], [0.01,0.02,0.03,0.04])

JSPEC.PlotRaw — Method

PlotRaw(ds:Dict; xlbl="Channels", ylbl="Counts", tlbl=ds.Name, verbose=true)::Figure

Draw a plot of the raw input data.

Arguments

ds JSPEC data set dictionary.

xlbl x-axis label.

ylbl y-axis label.

tlbl plot title.

verbose enables warning messages.

Examples

figraw = PlotRaw(newdataset)

JSPEC.PlotRebinned — Method

PlotRebinned(ds:Dict; xlbl="Channels", ylbl="Counts", tlbl=ds.Name, verbose=true)::Figure

Draw a plot of the rebinned input data.

Arguments

ds JSPEC data set dictionary.

xlbl x-axis label.

ylbl y-axis label.

tlbl plot title.

verbose enables warning messages.

Examples

figreb = PlotRebinned(newdataset)

JSPEC.RebinAncillaryData — Method

RebinAncillaryData(ds::Dict; verbose=true)

Rebin ancillary data (channels, channel energy, etc.) with the rebin schema identified for input data.

Arguments

ds JSPEC data set disctionary.

verbose enables warning messages.

Examples

RebinAncilaryData(newdataset)

JSPEC.RebinData — Method

RebinData(ds::Dict;minSN=5,verbose=true)

Rebin input data with a mininum S/N per bin.

Arguments

ds JSPEC data set disctionary.

minSN minimum S/N per bin.

verbose enables warning messages.

Examples

RebinData(newdataset)